+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.0 - 2020-01-02
+
+- Add rotation cycle to the FileAppender
+- Add custom appender feature
+- Add unit tests
+
 ## 0.1.0 - 2019-12-21
 
 - Initial release